Refactor cli module into submodules
Each submodule defines its cli interface.
